Updated to v13.2
- I also removed the org.secc.PayFlow.Reporting.dll since it was causing Rock to crash during startup.

+++ RockWeb/Blocks/WorkFlow/WorkflowTypeDetail.ascx.cs
@@ -323,15 +323,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                 LoadStateDetails( workflowType, rockContext );
 
                 // clone the workflow type
-                var newWorkflowType = workflowType.Clone( false );
-                newWorkflowType.CreatedByPersonAlias = null;
-                newWorkflowType.CreatedByPersonAliasId = null;
-                newWorkflowType.CreatedDateTime = RockDateTime.Now;
-                newWorkflowType.ModifiedByPersonAlias = null;
-                newWorkflowType.ModifiedByPersonAliasId = null;
-                newWorkflowType.ModifiedDateTime = RockDateTime.Now;
-                newWorkflowType.Id = 0;
-                newWorkflowType.Guid = Guid.NewGuid();
+                var newWorkflowType = workflowType.CloneWithoutIdentity();
                 newWorkflowType.IsSystem = false;
                 newWorkflowType.Name = workflowType.Name + " - Copy";
 
@@ -626,6 +618,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                 workflowType.ProcessingIntervalSeconds = null;
             }
 
+            workflowType.MaxWorkflowAgeDays = tbMaximumWorkflowAge.Text.AsIntegerOrNull();
             workflowType.LogRetentionPeriod = tbLogRetention.Text.AsIntegerOrNull();
             workflowType.CompletedWorkflowRetentionPeriod = tbCompletedRetention.Text.AsIntegerOrNull();
             workflowType.IsPersisted = cbIsPersisted.Checked;
@@ -745,6 +738,8 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
             }
             rockContext.SaveChanges();
 
+            var formBuilderEntityTypeId = EntityTypeCache.GetId<Rock.Workflow.Action.FormBuilder>();
+
             // add or update WorkflowActivityTypes(and Actions) that are assigned in the UI
             int workflowActivityTypeOrder = 0;
             foreach ( var editorWorkflowActivityType in ActivityTypesState )
@@ -772,12 +767,6 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                     SaveAttributes( new WorkflowActivity().TypeId, "ActivityTypeId", workflowActivityType.Id.ToString(), ActivityAttributesState[workflowActivityType.Guid], rockContext );
                 }
 
-                // Because the SaveAttributes above may have flushed the cached entity attribute cache, and it would get loaded again with
-                // a different context, manually reload the cache now with our context to prevent a database lock conflict (when database is 
-                // configured without snapshot isolation turned on)
-                EntityAttributesCache.Remove();
-                EntityAttributesCache.Get( rockContext );
-
                 int workflowActionTypeOrder = 0;
                 foreach ( var editorWorkflowActionType in editorWorkflowActivityType.ActionTypes )
                 {
@@ -800,14 +789,16 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                     workflowActionType.AttributeValues = editorWorkflowActionType.AttributeValues;
                     workflowActionType.Order = workflowActionTypeOrder++;
 
-                    if ( workflowActionType.WorkflowForm != null && editorWorkflowActionType.WorkflowForm == null )
+                    var isFormBuilderAction = workflowActionType.EntityTypeId == formBuilderEntityTypeId;
+
+                    if ( !isFormBuilderAction && workflowActionType.WorkflowForm != null && editorWorkflowActionType.WorkflowForm == null )
                     {
                         // Form removed
                         workflowFormService.Delete( workflowActionType.WorkflowForm );
                         workflowActionType.WorkflowForm = null;
                     }
 
-                    if ( editorWorkflowActionType.WorkflowForm != null )
+                    if ( !isFormBuilderAction && editorWorkflowActionType.WorkflowForm != null )
                     {
                         if ( workflowActionType.WorkflowForm == null )
                         {
@@ -1456,6 +1447,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                 tbProcessingInterval.Text = string.Empty;
             }
 
+            tbMaximumWorkflowAge.Text = workflowType.MaxWorkflowAgeDays.ToStringSafe();
             tbLogRetention.Text = workflowType.LogRetentionPeriod.ToStringSafe();
             tbCompletedRetention.Text = workflowType.CompletedWorkflowRetentionPeriod.ToStringSafe();
             cbIsPersisted.Checked = workflowType.IsPersisted;
